Fixed share indicator delay
